+++ pkgs/test_core/CHANGELOG.md
@@ -1,6 +1,9 @@
-## 0.2.16-dev
+## 0.2.16
 
 * Internal cleanup.
+* Add `customHtmlTemplateFile` configuration option to allow sharing an
+  html template between tests
+* Depend on the latest `test_api`.
 
 ## 0.2.15